move source files to src\ folder
