--- .qmake.conf
@@ -1,5 +1,0 @@
-load(qt_build_config)
-CONFIG += warning_clean
-DEFINES += QT_FOREACH
-
-MODULE_VERSION = 5.13.0
